start refactoring and cleanup of discovery module by creating dedicated subpackage
