+++ tests/test_discovery.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 import pytest
-from upnpavcontrol.core import discover
-
+from upnpavcontrol.core.discovery.registry import DiscoveryEventType
+from upnpavcontrol.core.discovery import utils
 from unittest.mock import Mock
 
 
@@ -16,9 +16,8 @@ async def test_discover_by_alive(started_mocked_device_registry):
 
     renderer_ssdp_data = await registry.trigger_renderer_alive()
 
-    discovery_callback.assert_called_once_with(
-        discover.DiscoveryEventType.NEW_DEVICE,
-        discover.udn_from_usn(renderer_ssdp_data['USN'], renderer_ssdp_data['NT']))
+    discovery_callback.assert_called_once_with(DiscoveryEventType.NEW_DEVICE,
+                                               utils.udn_from_usn(renderer_ssdp_data['USN'], renderer_ssdp_data['NT']))
 
     assert len(registry._av_devices) == 1
     assert len(registry.mediarenderers) == 1
@@ -33,8 +32,8 @@ async def test_discover_by_alive(started_mocked_device_registry):
     assert len(registry.mediaservers) == 1
 
     assert discovery_callback.call_count == 2
-    discovery_callback.assert_called_with(discover.DiscoveryEventType.NEW_DEVICE,
-                                          discover.udn_from_usn(server_ssdp_data['USN'], server_ssdp_data['NT']))
+    discovery_callback.assert_called_with(DiscoveryEventType.NEW_DEVICE,
+                                          utils.udn_from_usn(server_ssdp_data['USN'], server_ssdp_data['NT']))
 
     # alive message from an already known device should do nothing
     await registry.trigger_server_alive()
@@ -84,9 +83,8 @@ async def test_remove_by_byebye(mocked_device_registry):
     assert len(registry.mediarenderers) == 0
     assert len(registry.mediaservers) == 1
 
-    discovery_callback.assert_called_once_with(
-        discover.DiscoveryEventType.DEVICE_LOST,
-        discover.udn_from_usn(renderer_ssdp_data['USN'], renderer_ssdp_data['NT']))
+    discovery_callback.assert_called_once_with(DiscoveryEventType.DEVICE_LOST,
+                                               utils.udn_from_usn(renderer_ssdp_data['USN'], renderer_ssdp_data['NT']))
 
     await registry.async_stop()
 
@@ -97,6 +95,7 @@ async def test_scan_av_devices(mock_scanned_devices, started_mocked_device_regis
     discovery_callback = Mock(name='registry_event_callback')
     registry.set_event_callback(discovery_callback)
 
+    # ensure all outstanding (i.e. simulated) discovery events have been processed
     await registry._event_queue.join()
 
     assert len(registry._av_devices) == 2
