+++ tests/test_event_bus.py
@@ -1,5 +1,5 @@
 import pytest
-from upnpavcontrol.core import discover
+from upnpavcontrol.core.discovery import utils
 
 
 @pytest.mark.asyncio
@@ -16,6 +16,6 @@ async def test_device_found_event(webapi_client, mocked_device_registry):
             event1 = await websocket1.receive_json()
             event2 = await websocket2.receive_json()
             assert event1['event_type'] == 'NEW_DEVICE'
-            assert event1['udn'] == discover.udn_from_usn(renderer_ssdp['USN'], renderer_ssdp['NT'])
+            assert event1['udn'] == utils.udn_from_usn(renderer_ssdp['USN'], renderer_ssdp['NT'])
             assert event2['event_type'] == 'NEW_DEVICE'
-            assert event2['udn'] == discover.udn_from_usn(renderer_ssdp['USN'], renderer_ssdp['NT'])
+            assert event2['udn'] == utils.udn_from_usn(renderer_ssdp['USN'], renderer_ssdp['NT'])
