+++ upnpavcontrol/core/discovery/__init__.py
@@ -1,0 +1,3 @@
+from .registry import DeviceRegistry
+
+__all__ = ['DeviceRegistry']
